servod: fix ftdi logic
The previous version assumed that if the interface is a string and not a
dict, it's ftdi or dummy - turns out that's a bad assumption as v3 also
has string-only interfaces. This CL fixes that.

BUG=chromium:966954 //kind of
TEST=manual testing on v2, comes up as expected, and v3, seems to work
as well i.e. EC and AP consoles work

Change-Id: I4893a836a38239f025e3d82aecc05d4b8d731c65
Reviewed-on: https://chromium-review.googlesource.com/1666762
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Brian Nemec <bnemec@chromium.org>