ec: Wait the button release after issuing the async command
There are 2 implementations of emulating a button press.

On servo v2 or servo v4 with servo micro, it toggles the power button
line controlled by servo GPIO. It returns after the button release.

On servo v4 with CCD, it sends EC console command "powerbtn" with a
time interval that emulates the internal power button state change.
This console command is async. It returns immediately before the
button release. To align with the first implementation, need to wait
the given time interval for the button release.

BUG=b:76210256
TEST=Ran "dut-control pwr_button_hold:10000", which took 10s to return.
TEST=Ran "dut-control pwr_button_hold:1900", which took 1.9s to return.

Change-Id: I9b46d88f1a32c486947ecbd07a576cbf9920df83
Signed-off-by: Wai-Hong Tam <waihong@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/1667642
Tested-by: Caveh Jalali <caveh@google.com>
Reviewed-by: Shelley Chen <shchen@chromium.org>
Reviewed-by: Bob Moragues <moragues@chromium.org>
Reviewed-by: Matthew Blecker <matthewb@chromium.org>
Reviewed-by: Caveh Jalali <caveh@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>