servo v3: move atmega to uart0 in naming
This change is to avoid an issue where if cr50.xml is imported on a v3,
the atmega uart gets clobbered and we get all sorts of issues.
This moves the name to uart0_ to avoid that collision.

The real fix should be:
- ccd_cr50.xml has its own uart4/cr50 uart
- servo_micro.xml has its own uart4/cr50 uart
- there should be a uart4_pty control in cr50.xml ONLY for servo v2 that
conditionally turns on uart4 on v2.

As that requires a bit more plumming, we'll punt that to the new servod
and have this fix for now.

BUG=chromium:969709, chromium:976536
TEST=manual testing
sudo servod -b flapjack // on v3
localhost ~ # dut-control init_usb_keyboard:off
localhost ~ # dut-control init_usb_keyboard:on
2019-06-20 18:48:26,939 - kbHandlerInit - INFO - No device path
specified for usbkm232 handler. Use the servo atmega chip to handle.
2019-06-20 18:48:28,375 - kbHandlerInit - INFO - USBKM232: /dev/ttyO4

Change-Id: Ic22c592b4d7b74e30cc63054a845b28efca626ec
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1669197
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>