servo: add event object for each device
Use a device specific event object to track whether the device is
disconnected. Before they all used the same event object, so if CCD
disconnected all servo controls would wait not just ccd controls. For an
interface to wait for a device it needs to return the vid, pid, and
serialname in get_device_info. If the interface doesn't have
get_device_info, then the control will not wait for the interface to be
available.

This change only adds get_device_info to CCD interfaces, because those
are the only ones that can be reinitialized. Servo will die immediately
if any other device is disconnected.

BUG=b:137379310
BRANCH=none
TEST=None

Change-Id: I0b09d8af3be25e803b18fd66540a0d1f367f9aed
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1702807
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>