servo-logging: isolate log-dir logic
This change isolates log-dir logic from basiConfig logic when log-dir is
not defined. This is to allow different loggers to individually (e.g.
like for the bug here, ec3po loggers) to switch to DEBUG and have their
messages still show up in the stdout, while the overall system is still
in info. If log-dir is defined, servod uses specific handlers for stdout
and file handling, and then loglevel controls for specific loggers are
no longer supported.

BUG=chromium:993212
TEST=sudo servod -b nocturne & // observe normal logs
     dut-control ec_loglevel:debug // observe EC messages only show up
TEST=sudo servod -b nocturne --log-dir // observe normal logs
     dut-control ec_loglevel:debug // observe no EC messages show up as
     they're all already available in the log-file output (always debug)

Change-Id: I7160c857bd173ae82dfa902b82c3e547df6af8d1
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/1752591
Reviewed-by: Mary Ruthven <mruthven@chromium.org>