servo_mfg: upgrade mfg_servo_v4 script
This is an update to the existing script to work for v4 validation and
manufacturing using a few improvements

- The kernel code necessary to reflash the mac address will be natively
in Chrome OS images, and therefore will not require to load modules, or any of
that logic

- The previous logic hard-codes eth0 as the mac interface, which might
not be true. This version now uses the VID/PID of the ethernet dongle to
find the right usb device and the mac address from there

- After rewriting the serialname, the /serial file on the system is no
longer accurate and needs to reenumerate. This is necessary for the
validation script to work to ensure that the rewrite was successful. For
this, the mechanism is to toggle authorized on the host hub that the
v4's stm32 sits on.

BUG=chromium:931904
TEST=mfg_servo_v4 -m $MAC -s $SERIAL --no_flash // with mismatched mac
TEST=mfg_servo_v4 -m $MAC -s $SERIAL --no_flash // with correct mac
TEST=mfg_servo_v4 -m $MAC -s $SERIAL --no_flash // with new serial
     cat /path/to/v4/serial // observe new serial showing in sysfs

Change-Id: I6fec8c93301571def35ac23a3bb3a22ef166d3ec
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/1761093
Reviewed-by: Todd Broch <tbroch@chromium.org>