servo: add version number
Add a version number that incorporates major features and labstation
pushes, so it's easier to track labstation pushes and features from
servod --version. Right now you need to checkout the hash and look at
hdctools to figure out what features are in the version. Hopefully with
this new versioning, this will be a little more obvious. It will be up
to committers to update the MAJOR_FEATURE_FIELD if they want their
feature tracked.

BUG=none
BRANCH=none
TEST=servod --version

Change-Id: If9ed26068ac15a762fb413074dd6a5eb26131bd5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/1772222
Tested-by: Mary Ruthven <mruthven@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>