servo: hub support for v4 image usb stick
This change allows users (and the lab) to attach a usb hub onto the
type-A port usually reserved for the usb stick. The caveat is that the
stick has to be attached to that hub's first port (as they enumerate in
sysfs). This means
- for each hub that port has to be identified to use this properly
- some larger hubs (that are internally a chained hub) are excluded as
they use the first port of the first hub to chain the second hub.

BUG=b:140964613
TEST=sudo servod -b atlas (v4 with hub, stick on port 4)
     dut-control image_usbkey_dev

image_usbkey_dev:

TEST=sudo servod -b atlas (v4 with hub, stick on port 1)
     dut-control image_usbkey_dev

image_usbkey_dev:/dev/sda

TEST=sudo servod -b atlas (v4 without hub)
     dut-control image_usbkey_dev

image_usbkey_dev:/dev/sda

Change-Id: I171e50331e7f6deee2f21f89ff32002061d1bb8f
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/1860040
Reviewed-by: Todd Broch <tbroch@chromium.org>