active_v4_device: check if ccd TX is enabled
The +TX in the state flags shows if ccd will interfere with servo. It
doesn't really matter what state cr50 thinks servo is in. It matters if
TX is enabled and will interfere with servo. Modify using_ccd to check
if +TX is in the state flags to determine if we're using ccd.

BUG=chromium:1016842
BRANCH=none
TEST=change the active v4 device.

Change-Id: I177b5717c0e192898debf6802371a02cdca2f6d0
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/1874739
Reviewed-by: Namyoon Woo <namyoon@chromium.org>