servo_micro: add dut control for holding USART low
To put some ECs into UART (UUT) programming mode, the EC_TX_SERVO_RX
line needs to be held low by the servo when the EC is resetting. The
console command has been added in chromium:1884190, and this is the
dut-control interface that the flash_ec script will use.

BRANCH=none
BUG=b:143163043
TEST=flashed phaser using UUT with reworked servo_micro

Cq-Depend:chromium:1884190
Change-Id: I95950b9768a3224763ced4886d2dc8b284333b91
Signed-off-by: Jett Rink <jettrink@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/1884410
Reviewed-by: Edward Hill <ecgh@chromium.org>
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>