servo_mfg: use bin files from updater
This change is to avoid packaging seperate .bin files with servo_mfg, but rather to just use the servo_updater bin
files. This ensures that whenever those get uprev'd, the mfg script also
has up-to-date files.

This means that the binary files shipped with servo_mfg are the ones
only for mfg - which for now is just the keyboard.hex

BUG=b:142143881
TEST=mfg_servo_v4 -s $serial --force_atmega -m xx:xx:xx:xx:xx:xx

// observe full update process happens successfully

Change-Id: I44635caef36086e4cc60d4dd4616544e4c959203
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/1885387
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>