rammus: route warm_reset control to cold_reset
on rammus it seems that warm_reset does nothing at least, it does not
warm_reset the DUTs. This causes failures in the lab.
This solution is similar to what's done on kukui - where warm_reset does
not exist - and we route warm_reset to be really cold_reset under the
hood.

BUG=chromium:1018322
TEST=sudo servod -b rammus &
     dut-control -i warm_reset:
warm_reset:cold_reset

TEST=sudo servod -b rammus &
     dut-control warm_reset:on warm_reset:off

>> observe DUT rebooting

Change-Id: I594c8babdf1208057837430ffd9670e4bcda91c2
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/1912630
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>