usb_image_manager: improve robustness and error reporting
The main point of this change is to improve robustness and error
reporting on the usb hub. To do this, the change has added more
descriptive error messages that both explain what locations in sysfs
were tried (and failed) and where the stick should be plugged in (in
case the user plugged it into the wrong (out of many) usb port.

BUG=None
TEST=sudo servod -b krane // v4
     dut-control image_usbkey_dev //image usb port has stick
> image_usbkey_dev:/dev/sda

     dut-control image_usbkey_dev //image usb port empty
> [...]WARNING - No USB storage device found for image transfer. Make sure you
are using the usb port next to the ethernet adapter.
> [...]WARNING - Stick should be at /sys/bus/usb/devices/1-3.3.4.
> [...]WARNING - If using a hub on the image key port, please make sure to use
port 1 on the hub. This should be at /sys/bus/usb/devices/1-3.3.4.1.

     dut-control image_usbkey_dev //image usb port with hub & stick
> image_usbkey_dev:/dev/sda

     dut-control image_usbkey_dev //image usb port with empty hub
> [...]WARNING - No USB storage device found for image transfer. Make sure you
are using the usb port next to the ethernet adapter.
> [...]WARNING - Stick should be at /sys/bus/usb/devices/1-3.3.4.
> [...]WARNING - If using a hub on the image key port, please make sure to use
port 1 on the hub. This should be at /sys/bus/usb/devices/1-3.3.4.1.

Change-Id: I064f8388885ec8f79329594bb7b2f0e365b980ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/1913479
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>