servod: improve error reporting on non-available ports
The current implementation pre-populates sopts.port and therefore, no
matter if a specific port was requested, or if simply no open port was
found in the range specified, servod will report the same error message.
This change improves on that reporting by checking whether
start_port==end_port. This condition is only caused by specifically
requesting a port. Otherwise the range is larger.

BUG=chromium:1029065
TEST=sudo servod -p 9998 &
     sudo servod -p 9999 &
     sudo servod -p 9998
> CRITICAL - Port 9999 is busy

TEST=//change default range to 9998 to 9999
     sudo servod -p 9998 &
     sudo servod -p 9999 &
     sudo servod
> CRITICAL - Could not find a free port in 9998..9999 range

Change-Id: Icafb9cef45d2b97c8da0a8a71a13f084bf4d2b0a
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/1956334
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>