servo_kb: throw a warning, not a giant error on missing usb kb
Some boards (boxes, tablets) require usb keyboard emulation for the main
keyboard on servod. Some servos, like just servo micro, do not have that
functionality. Mix those two, and you can't use the keyboard controls on
servod e.g. dut-control ctrl_d:tab.

This might be working as intended however, if the user is not intending
to use those behaviors. This change stops throwing a giant error and
just prints a warning, letter the user know the situation.

BUG=chromium:1028029
TEST=manual testing
     sudo servod -b krane // only servo micro

2019-12-09 15:19:20,267 - kbHandlerInit - WARNING - The servo setup does
not have a usb keyboard emulator. Will not throw an error, but note that
the keyboard controls will fail, as no keyboard could be setup.

Change-Id: I6eb15697d317467588f530b9309bdb274af9de16
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/1959144
Reviewed-by: Fei Shao <fshao@chromium.org>