servo_keyboard: split real keyboard and power key drivers
For some platforms, the keys are external (usb kb). The power-key
controls however are internal: EC, Cr50, or the servo micro power button
driver.
This change splits the keyboard handling so that there is a different
driver for main keys vs power keys.
The main motivation: if a user is playing with a usb keyboard DUT with a
servo that does not support usb keyboard (micro, ccd only), they should
not be deprived of the power key functionality. This change fixes that.

BUG=chromium:1028029
TEST=sudo servod -b krane // ccd only, suzy-q

     dut-control power_key:tab
> observe power key being pressed
     dut-control ctrl_d:tab
Problem with ['ctrl_key:tab'] :: Keyboard handler not setup.

TEST=sudo servod -b krane // servo micro only

     dut-control power_key:tab
> observe power key being pressed
     dut-control ctrl_d:tab
Problem with ['ctrl_key:tab'] :: Keyboard handler not setup.

Change-Id: I61b1c561eb186e9f25a21f9bdb21d858b75d6775
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/1959406
Reviewed-by: Todd Broch <tbroch@chromium.org>