servo keyboard: introduce base template and noop handler
To properly accomodate the lack of usb keyboard on ccd/micro only and
boards (e.g. krane) that require a usb keyboard as a main keyboard, this
change adds a NoopHandler that simply always reports that the keyboard
handler is turned off, and prints warnings whenever the user tries to
turn on/off that handler.

This has two advantages:
- whenever the user/some autotest code by default tries to enable the
handler, only a warning is printed indicating that real functionality is
absent
- when the user/a test tries to use a real key on that noop keyboard,
then a warning will be thrown. However, the logs make it easy to trace
back that the real keyboard doens't exist.

Essentially we delay the error reporting until there is something we
actually cannot do: like pressing a real key, and not on keyboard
initialization.

BUG=chromium:1028029
TEST=sudo servod -b krane // micro only: testing noop handler
     dut-control init_keyboard:on
> 2019-12-17 12:16:48,578 - NoopHandler - WARNING - Using a noop keyboard
  handler. Check logs to see why it is in use and address issue if full
  keyboard functionality is needed.
     dut-control init_keyboard:off
> 2019-12-17 12:16:39,862 - NoopHandler - WARNING - Using a noop keyboard
  handler. Check logs to see why it is in use and address issue if full
  keyboard functionality is needed.

     dut-control init_keyboard
> init_keyboard:off

TEST=sudo servod -b krane // micro + v4: testing EC handler
     dut-control init_keyboard
> init_keyboard:off
     dut-control init_usb_keyboard:on
     dut-control init_keyboard
> init_keyboard:on

TEST=sudo servo -b atlas // micro + v4: testing UsbKb handler
     dut-control init_keyboard
> init_keyboard:on
     dut-control init_usb_keyboard
> init_usb_keyboard:off
     dut-control init_usb_keyboard:on
     dut-control init_usb_keyboard
> init_usb_keyboard:on

Change-Id: I6eee01c2f81b7ceb7d9f9ec48f058a81cfd77007
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/1972519
Reviewed-by: Fei Shao <fshao@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>