c2d2: add ap_flash_select control
In order to power the SPI flash on the DUT, we need to use the H1
console to enable the AP_FLASH_SELECT gpio. Add a new dut-control for
this to make flashing the AP easier.

Also correct a bug in get_ec_mode. The get version of this dut-control
has probably not been used in a very long time. The regex and matching
index both needed to be updated to make it work.

BRANCH=none
BUG=b:145314772
TEST=dut-control ap_flash_select gets and sets correctly with newest H1
image

Change-Id: I9e9dc5b9c68da360c5188c9c340833d8adebfc39
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/1995612
Tested-by: Jett Rink <jettrink@chromium.org>
Reviewed-by: Diana Z <dzigterman@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Commit-Queue: Jett Rink <jettrink@chromium.org>
Auto-Submit: Jett Rink <jettrink@chromium.org>