usb_image_manager: turn on power on image_usbkey_[dev/direction]
image_usbkey_[dev/direction] are intended to be higher level
abstractions that make sure that everything is in order to use a usb
stick to flash an image/boot from it. This change fixes an issue where
if the stick was already facing the right direction i.e. DUT and DUT was
being requested, but the power to the usb port was turned off, the power
to the port would remain turned off. That's counter to how it works if
the direction was mismatched, and counter to the goal of these
abstractions.

This change now enforces that calls to image_usbkey_[dev/direction]
always end up with the usb port being powered.

Note that if that is not desired, the user still has the freedom to use
image_usbkey_mux (an unsafe direct control to the mux that does not
power cycle during direction changes) to manually manage this.

BUG=chromium:1017308
TEST=manual test for all 4 possible start conditions.
     Note that rotate_servod_logs:yes is added as having no 'set'
     control in dut-control allows the caller to reshuffle the control
     order. However , here we'll want to capture the previous state
     first, before showing that the usb stick is properly facing the
     host and enumerated at the end. That call is omitted below.
Case 1: powered, facing DUT
     dut-control image_usbkey_pwr image_usbkey_direction image_usbkey_dev
     image_usbkey_pwr:on
     image_usbkey_direction:dut_sees_usbkey
     image_usbkey_dev:/dev/sda
Case 2: unpowered, facing DUT
     dut-control image_usbkey_pwr image_usbkey_direction image_usbkey_dev
     image_usbkey_pwr:off
     image_usbkey_direction:dut_sees_usbkey
     image_usbkey_dev:/dev/sda
Case 3: powered, facing SERVO HOST
     dut-control image_usbkey_pwr image_usbkey_direction image_usbkey_dev
     image_usbkey_pwr:on
     image_usbkey_direction:servo_sees_usbkey
     image_usbkey_dev:/dev/sda
Case 4: unpowered, facing SERVO HOST
     dut-control image_usbkey_pwr image_usbkey_direction image_usbkey_dev
     image_usbkey_pwr:off
     image_usbkey_direction:servo_sees_usbkey
     image_usbkey_dev:/dev/sda

Change-Id: I7db0bf3ce34dc4ee3d66d216206fd2ccbb1db14b
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2017952
Reviewed-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Brian Nemec <bnemec@chromium.org>