servodtool: add tool infrastructure
The motivation here is that tools and utilities can be built seperately,
have individual unit-tests, and mix-and-match easier. A new command-line
tool then simply needs to conform to the interface, and register itself
it tools/__init__.py and servodtool will support it.

The first tool (what used to be the command-line tool for servodutil) is
now:
servodtool instance

servodtool instance --help
usage: servodtool instance [-h] {show,stop,rebuild,show-all}

proving the same functionality and sub-commands that servodutil provided before.

This change adds the final piece of the new util/tool framework, by
- creating a tool interface for future command-line tools
- removing servodutil
- replacing its command-line functionality with the instance tool
- creating servodtool as the router for all servod command line tools
- adding a legacy command-line tool servodutil that maps to servodtool
instance

BUG=chromium:1052541
TEST=servodtool instance show-all
> port : 9999
> serials : [u'C1804020019', u'SNCQ00098']
> pid : 73373

TEST=servodtool instance show -p 9999
> port : 9999
> serials : [u'C1804020019', u'SNCQ00098']
> pid : 73373

TEST=servodtool instance show -p 9998
> No servod scratch entry found under id: 9998.

TEST=sudo rm /tmp/servoscratch/*
     servodtool instance rebuild -p 9999
> Port 9999 not registered but has a servod instance bound to it.
> Retroactively adding the instance.

TEST=sudo servodtool instance stop -p 9999
> Servod instance associated with id 9999 turned down.

Change-Id: I991d10d11f689372ab028b680ecd30a7579a3f21
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2066043
Reviewed-by: Brian Nemec <bnemec@chromium.org>