servo: Refactor servo set/get logging, at INFO level.
This change combines two earlier prototype CLs, and separates the log
formatting logic into wrappers called by the actual set/get logic.

When running tests, it's nice to be able to see all the servod calls
the test is making, without having to open the DEBUG log full of
console output and implementation details.

When adding the logging, I discovered that some set() and get() calls
result in additional calls, so now those are logged with indentation.

Excessively long results of commands (such as ec_uart_stream) are now
abbreviated, to reduce noise on the terminal or info log.

When encountering exceptions, the traceback previously wasn't logged,
making it hard to tell what the exception was, or where it came from.
Now the type is logged at the ERROR level, and the full traceback is
logged at DEBUG level.

BUG=None
TEST=Run a faft suite, check for logging of commands.
TEST=Run firmware_ConsecutiveBoot, check for logging of ptyError.

Change-Id: I877d05733e1d13f84152719f040b652f787810a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2068944
Tested-by: Dana Goyette <dgoyette@chromium.org>
Commit-Queue: Dana Goyette <dgoyette@chromium.org>
Reviewed-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>