servodtool: leverage tool.Error in instance
The previous CL introduced a common Error method in tool to make sure
that the tools exit with a non-zero exit code, as they are command-line
tools. This change leverages that call in instance as well.

This CL also leverages the Error() call to provide information to the
user when a subparser implements a command, but the command itself
cannot be found.

BUG=chromium:1052541
TEST=servodtool instance fake
> Tool does not recognize command 'fake'. It should be implemented as a
> method called 'Fake'.
     echo $?
> 1

Change-Id: I127a29059b39a29d7001749b08ec11686a3a77a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2077117
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Brian Nemec <bnemec@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>