servo_logging: Adjust controls logging alignment.
Now that the set/get messages are only in the DEBUG level, the
indentation is misalgned due to the function name in the log format.
Adding a filter allows fixing it without changing the logging format.

Additionally, the indentation is now done via format string instead
of a method.

BUG=chromium:1057830
TEST=Run servod, grep DEBUG log for Controls

Change-Id: I3381e2807ff39c32a028c65934b2139a0ad636dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2085475
Tested-by: Dana Goyette <dgoyette@chromium.org>
Commit-Queue: Dana Goyette <dgoyette@chromium.org>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>