servod: unified usage message
The servod invocation uses two parsers, one for server arguments, one
for device arguments under the hood. The first one parsers out all the
server args and then hands off the remainder to the second one. The
issue here is that if an option is 'unknown', then the second one will
only prints its own usage. This change fixes that by pointing both
'usage' printing logics to the unified 'help_displayer' parser, the one
that has the union of the arguments, and we use to print meaningful help
messages.

BUG=None
TEST=sudo servod --nonsense
// observe now all options, including the server ones, being shown

Change-Id: Ie692ad239e6a70d3364a2122d17d8cf595a38661
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2113731
Reviewed-by: Brian Nemec <bnemec@chromium.org>
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>