servod: add sversion logic
This change borrows heavily from the EC getversion.sh logic to generate
a meaningful servod version string. Essentially, on build time we
dynamically generate a python dictionary with all the information to
then generate version string(s) that are (1) python compliant (2)
informative. The logic is to use Tags so for the major/minor number, and
then have the number of commits since the tag as the third number.
Like in the ec case, we use the last hash to help anchor the change, and
provide + and - to indicate a dirty or clean repo.
Since there are some limitations on what the version string can be in a
compliant python package (no hex hash for example),
we simply append .dev to the version if hdctools is built locally to
differentiate it.

The next change will introduce an 'extended' version that can be queried
that has all the information to facilitate debugging.

BUG=chromium:1063519
TEST=./getversion.sh
     // inspect sversion.py

TEST=cros tryjob -g 2110752 fizz-labstation-release-tryjob
    // inspect that in a 'builder', the git history is available,
    // and the version string meaningful

Change-Id: Ic43fa3ce4f55c04d1fc0ae8c3e7429dd0fc606e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2113732
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>