fw_wp_state: Add option to make WP open-drain
In the original approach, the WP output is through a voltage level
shifter. The reference voltage of the output is spi2_vref (or
fw_wp_vref). In some cases, we don't know to supply the spi2_vref,
like preventing current leak when AP is off. It makes the level
shifter not active.

This CL supports an option to make the WP an open-drain output and
not use the level shifter.

BUG=b:151058089
TEST=With a follower CL, tested switching the fw_wp_state to on
and off working reliably in Trogdor, even AP is off.

Change-Id: I96432cbe986a97f25fac4bc0a452674b23a201fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2119955
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@google.com>