setup: pull ina building into own method
We are doing multiple steps in the build_py now with sversion. To
prepare for that, refactor this by pulling the ina building out into its
own function. As this is part of moving to python3, also change the
import structure in the ina generation file so that the helper modules
are always found: in setup.py and through the command-line.

BUG=chromium:1063519
TEST=sudo emerge hdctools -j32

Change-Id: I432290ca31a94205913d77e80482c84d90fb4055
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2125084
Reviewed-by: Todd Broch <tbroch@chromium.org>
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>