Trogdor: Use Cr50 UART approach to modify WP state
The Cr50 UART approach is to access the Cr50 UART console to call the
command to overwrite the write protection state. Usually use this
approach for the CCD setup.

When using the servo v2 (or similar servo through the servoflex)
setup, it usually supplies power to spi2_vref and then overdrives the
WP signal. However, this results current leak on Trogdor.

This CL forces to use the Cr50 UART approach even on the servo v2
setup.

This requires accessing Cr50 UART. For the boards, in which the Cr50
UART is repurposed as JTAG, it is intended not working. Some error
like "cr50 uart is unresponsive" will be returned. Have to:
 * use the CCD setup instead
 * rework the board to make Cr50 UART available

BUG=b:151058089
TEST=Used the CCD setup; setting fw_wp_state worked.
TEST=Used the servo_micro setup; setting fw_wp_state returned
"cr50 uart is unresponsive" error on the board without rework.

Change-Id: I0d71f60da140ed421859adf7b2c1bc77dc3c4702
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2130484
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@google.com>