pty_driver: detect read-only console
This change helps detect if a console that times out is a read-only
console or not.
This means that, in the scenario where the there is output on the
console (i.e. we can tell it's read only or not) the driver checks
whether the command was actually received on the console side. If not,
it's likely a read-only console.

BUG=b:153366900

// Error message below formatted from regular output

TEST=sudo servod -b volteer --allow-dual-v4 & // EC console on micro: rw
                                              // EC console on ccd: ro
     dut-control ec_uart_cmd:reboot sleep:1 ec_board // test micro
> ec_board:volteer

// reboot issued in both tests to stimulate EC output well beyond the
// control finishing

TEST=sudo servod -b volteer --allow-dual-v4 & // EC console on micro: rw
                                              // EC console on ccd: ro
     dut-control ec_uart_cmd:reboot sleep:1 ccd_cr50.ec_board // test ccd
Problem with ['ec_uart_cmd:reboot', 'sleep:1', 'ccd_cr50.ec_board'] :: ['ver'].
EC: There was output: (likely read_only console)
[0.924054 PB task 9 = recovery]
[0.924460 PB PCH pwrbtn=HIGH]
[0.924757 PB task 0 = idle, wait -1]
[1.010184 power state 2 = S3, in 0x0039]
[...]

Change-Id: Ie85ce087df437867525fb43897e5aa19bd75a1aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2151790
Reviewed-by: Dana Goyette <dgoyette@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>