stm32usb: introduce locking for read and write endpoints
When the stm32uart is reinitializing, there is a race between the
watchdog thread setting up the USB endpoints again, and the UART threads
trying to use those endpoints. In rare (but not impossible
circumstances) this leads to segfaults in pyusb or libusb, or other
unhandled python errors (handler is None, etc).

For other stm32 endpoints this is less of an issue, but still not
impossible. A control might sneak through the interface before the
watchdog can catch that the device has fallen off. In that case the
reinitialization of the watchdog is racing against the i2c/gpio control
on the server thread.

This change introduces a lock each for the read endpoint and write
endpoint. Those are per-interface, and thus live inside stm32usb - as
one instance represents one interface.

Helpers are introduced so that read and write to the ep's are
automatically guarded by the locks, and reinitialization has to first
make sure to acquire both locks, before proceeding.

Lastly, all lock acquisition is done with a 60s timeout. This is more
than enough in normal servod circumstances, but would prevent the system
from hanging forever if another currently unknown issue prevents the
reinitialization from releasing all tokens, or either UART threads from
releasing theirs.

BUG=b:153883605
TEST=sudo servod -b nocturne &
     while true; do for ctrl in "watchdog_remove:ccd" \
     "servo_v4_dts_mode:off" "servo_v4_dts_mode:on" "watchdog_add:ccd" \
     "ec_board" "cr50_vsion"; do dut-control $ctrl; sleep 0.01; done; done

ec_board:nocturne
cr50_version:0.4.16/cr50_v1.9308_B.559-683b45f
[...] for over an hour

     sudo ls /proc/191712/fd | wc -l
119 // this number can rise up to ~123, and then returns back to a
    // stable 119. This is verify that there is no regression with
    // leaking fds

Change-Id: I10189460017c9934aa7f7d1d5104a06830cbaf2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2161848
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Namyoon Woo <namyoon@chromium.org>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>