usb_image_manager: expand hub port logic for image_usbkey_dev
On some linux distribution it appears that at least servo v2's usb stick
enumerates as a different hub port: 2 vs the 6 that it does under
gLinux/mint linux.

This change is to accomodate a range of candidate locations, so that for
a given servo device, multiple locations can be checked to find the usb
stick.

Note: this is to facilitate compability across multiple linux
distributions, and not to faciliate using different/multiple usb ports
on the servo devices. The physical port is still the same one, and there
is still only one option for the physical port. All controls that center
around muxing that physical port or turning power on/off for that
physical port still only apply to the designated port.

BUG=b:156225711
TEST=dut-control image_usbkey_dev
image_usbkey_dev:/dev/sda

Change-Id: I44d7f412572c9937f9540a37ba953be23246e25b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/hdctools/+/2216573
Tested-by: Dossym Nurmukhanov <dossym@chromium.org>
Reviewed-by: Dossym Nurmukhanov <dossym@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>